+++ R/package_solitude.R
@@ -1,5 +1,8 @@
 #' @name solitude
 #' @title An Implementation of Isolation Forest
-#' @description Isolation forest is anomaly detection method introduced by the paper Isolation based Anomaly Detection (Liu, Ting and Zhou <doi:10.1145/2133360.2133363>)
+#' @description Isolation forest is an anomaly detection method introduced by the paper Isolation based Anomaly Detection (Liu, Ting and Zhou <doi:10.1145/2133360.2133363>)
 #' @author Srikanth Komala Sheshachala
+#' @importFrom data.table :=
 "_PACKAGE"
+
+`:=` = function(...) NULL